ExtraClasse/src/backend/requirements.txt
django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0